requirements.txt
FastAPI
SQLAlchemy
fastapi-sqlalchemy
uvicorn
numpy
python-nubia
termcolor
requests
